+++ CHANGES.md
@@ -10,7 +10,7 @@ v0.8.2 (unreleased)
 -------------------
 
 * Make sure histograms are updated if only the attribute changes and the
-  limits and number of bins stay the same.
+  limits and number of bins stay the same. [#1012]
 
 v0.8.1 (2016-05-25)
 -------------------