+++ CHANGES.md
@@ -4,6 +4,10 @@ Full changelog
 v0.8.2 (unreleased)
 -------------------
 
+* Fix a bug that caused multiple custom viewer classes to not work properly
+  if the user did not override ``_custom_functions`` (which was private).
+  [#810]
+
 * Make sure histograms are updated if only the attribute changes and the
   limits and number of bins stay the same. [#1012]